pyproject.toml
[tool.poetry]
name = "map-istanbul"
version = "0.1.0"
description = ""
authors = ["kemalbastak <bastakkemal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework-gis = "^1.1"
django-rest-knox = "^5.0.1"
pandas = "^2.2.2"
django-currentuser = "^0.8.0"
django-soft-delete = "^1.0.14"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
openpyxl = "^3.1.5"
django-debug-toolbar = "^4.4.6"
django-map-widgets = "^0.5.1"
asgiref = "^3.8.1"
attrs = "^24.2.0"
certifi = "^2024.8.30"
charset-normalizer = "^3.3.2"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
et-xmlfile = "^1.1.0"
idna = "^3.10"
inflection = "^0.5.1"
itypes = "^1.2.0"
jinja2 = "^3.1.4"
jsonschema = "^4.23.0"
jsonschema-specifications = "^2023.12.1"
markupsafe = "^2.1.5"
numpy = "^2.1.1"
openapi-codec = "^1.3.2"
pyjwt = "^2.9.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.2"
pyyaml = "^6.0.2"
referencing = "^0.35.1"
requests = "^2.32.3"
rpds-py = "^0.20.0"
simplejson = "^3.19.3"
six = "^1.16.0"
sqlparse = "^0.5.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.1"
uritemplate = "^4.1.1"
urllib3 = "^2.2.3"
django-import-export = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
